+++ src/z80asm/CHANGES.md
@@ -4,6 +4,13 @@ Z88DK Z80 Module Assembler Change Log
 2016
 ----
 
+- 2016-10-08
+
+  Add consolidated object file.
+  When -o objectfile is given without -b, all the object modules from the command line 
+  are merged into one consolidated object file. The module local symbols are renamed 
+  <module>_<symbol> to avoid clashes in the merged symbol table.
+
 - 2016-09-17
 
   Remove -M, --obj-ext option. The object file has always the .o extension.